Commented out the maintainer commands of the docker files as these are deprecated and really offer no benefit
also created the azure image code, this will create an image with base support for azure-cli. The idea is that I will use this as a base for other images to perform various functions using azure.

also added cron image to give my docker images the ability to use cron jobs